Add filter for preserve http.
This filter will actually be very handy.

It can be used to only preserve HTTP share counts for posts that we published before SSL was implemented, thus saving API calls for all new posts which do not need to look up and preserve share counts for HTTP.

With this filter you would take the post ID, look up the post published date, and see if it is before the date SSL was implemented. If it was, return true and HTTP counts will be preserved. If not, return false to save API calls and reduce overhead.